Ready to push for stateless PR
- Organized components
- moved app.js to containers
- factorized stuff
