+++ src/client/index.js
@@ -1,24 +1,17 @@
 import React from 'react'
 import ReactDom from 'react-dom'
+import createLogger from 'redux-logger'
+import { Provider } from 'react-redux'
+import { createStore, applyMiddleware } from 'redux'
+
+import App from './containers/App'
 
 // A man must have data
 import fake from './data'
 
-import Todo from './render'
-
-const App = React.createClass({
-  getInitialState: function() {
-    return { lists: this.props.data }
-  },
-  render: function() {
-    return (
-      <div className="app-wrapper">
-        <h1>A fantastic Todo is on its way !</h1>
-        <Todo lists={this.state.lists} />
-      </div>
-    )
-  }
-})
-
 // .. with a little help from his frameworks
-ReactDom.render(<App data={fake}/>, document.getElementById('react-wrapper'));
+ReactDom.render(
+  <Provider>
+    <App data={fake} />
+  </Provider>, document.getElementById('react-wrapper')
+);
